requirements.txt
discord.py
youtube-dl
pytube
aiohttp
ffmpeg-python